Generate tile of package logos

This function returns an interactive html tile view of the packages either listed in the packages option, or all of the loaded packages. When rendered interactively, the result is output in the viewer. When rendered in Quarto or RMarkdown, the tile becomes part of the rendered html. If local images are provided, only these images will be used, excluding loaded packages.

Usage

make_tile(
  packages = NULL,
  local_images = NULL,
  local_urls = NULL,
  dark_mode = FALSE
)

Arguments

packages

Character vector of package names to include (default: NULL, which uses loaded packages)

local_images

Optional character vector of local image paths to add to the tile

local_urls

Optional character vector of URLs for each of the local images passed

dark_mode

Draw the tile on a dark background?

Value

Path to the output file